+++ cobs/settings.hpp
@@ -10,11 +10,13 @@
 #define COBS_SETTINGS_HEADER
 
 #include <cstdlib>
+#include <stdint.h>
+
 
 namespace cobs {
 
 //! run COBS using parallel threads, default: all cores
-extern size_t gopt_threads;
+extern unsigned gopt_threads;
 
 //! whether to load the complete index to RAM for queries.
 extern bool gopt_load_complete_index;
